+++ apps/backoffice-v2/public/locales/en/translation.json
@@ -1,1 +1,3 @@
-{}
+{
+    "home.greeting": "Welcome"
+}
